application/requirements.txt
RPi.GPIO
Flask
pybluez
python-nmap
